requirements.txt
flask
requests
beautifulsoup4
python-dotenv
markdownify
